v1.00
Renekton - Crocodile Dundee v1.00 by QQQ

Changelog:
-- 0.01 --
- First BetaRelease
-- 0.02 --
- Fixed an error that caused huge FPS-Drops
-- 0.03 --
- Added a AutoIgnite function
- Fixed some Texterrors in the menu
- Added W-Animationcanceling with Tiamat/Hydra
a. Added to the "Extra"-Menu (Enabled by default)
-- 1.00 --
- Removed the LaneClearKey and the JungleClearKey
- Added a new ClearKey for both - Jungle- and LaneClearing (u need to
set the SOW, MMA, SAC LaneClearKey to the same as this key or you won't
AA)
- Changed Standardhotkeys to match SOW functions (you can still edit
them as you like)
- Changes to the "Extra"-menu:
a. Added a new Slider for the Animationcancelingdelay (standard +
recommended is 0.3)
- Changes to the "SBTW-Combo"-menu:
a. Added an option do enable/disable E2 (dice)
- Changes to the "Harass"-menu:
b. Added an option do enable/disable E2 (dice)
- Changes to the "LaneClear"-menu:
a. Added an option do enable/disable E2 (dice)
- Changes to the "JungleClear"-menu:
a. Added an option do enable/disable E2 (dice)
- Added the new Harassfunction E-Q-W-Tiamat-E back to startingPos
- Added a fps-lighter way for under-towercalculations
- Added the new smartKillStealfunction
a. Tries to KS with Q,W,E1 and E2
b. (still in testing - if it doesnt work for you change back to EQWE to
enemy)
- Bugfixes:
a. Fixed a bug with Tiamat/Hydra Animationcanceling while SBTW and
UseItemfunction. Should now reset properly.
b. Fixed and error with lichbane, the caused the script to print an
error/crashed the script