v0.15
-- 0.15 --
- Removed the LaneClearKey and the JungleClearKey
- Changed Standardhotkeys to match SOW functions (you can still edit
them as you like)
- Added a new ClearKey for both - Jungle- and LaneClearing (u need to
set the SOW, MMA, SAC LaneClearKey to the same as this key or you won't
AA)
- Bugfixes:
a. Fixed and error with lichbane, the caused the script to print an
error/crashed the script
b. Fixed an error in the Q-Castingfunction that prevented casting from
time to time