v1.01
-- 1.01 --
- Reworked the whole Autojump/Shield logic - should be a lot more
responsive now
- Bugfixes:
a. Fixed a bug that spammed an error if you enabled the drawing of the
W-Range
b. Fixed a bug that spammed an error and caused LaneClear not to work
with the inbuild orbwalker
c. Fixed a bug that spammed an error if AutoIgnite was enabled and
prevented to cast it on an enemy