v1.03
-- 1.03 --
- The script now detects if you are using MMA/SAC:R and tells you if it
got found or not
- The script automatically disables/enables the inbuild SOW-Orbwalker
depending on the Orbwalker
- The script now detects if you press a key in SAC:R
a. That means if you press a key in SAC:R you automatically activate the
right function in the Rengarscript
b. This applies to the MixedModeKey -> Harass, LaneClearKey -> Clear and
the Combokey -> SBTW
- Bugfixes:
a. Fixed a bug that prevented Q from casting sometimes