v2.04
-- 2.04 --
- Improved the Hitchance and Collision for spearollision for spears
- Changes to the Spells (Cougars
- Changes to the Spells (Cougar W):
a. Added a toggle to check the minimum distance to jump (standard =
enabled)
b. Added a slider to set the range that it needs to your target to jump
(0-350 - standard = 200)
- Bugfixes:
a. Fixed a bug where Nidalee switched from cougar to human even if the
manacheck was enabled and your mana was low